requirements.txt
amqp==5.0.2
asgiref==3.2.10
backcall==0.2.0
billiard==3.6.3.0
cachetools==4.1.1
celery==5.0.2
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cloudinary==1.22.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
decorator==4.4.2
detect==0.0.0
Django==3.1.1
django-celery-results==2.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.1
drf-yasg==1.20.0
google-api-core==1.23.0
google-api-python-client==1.7.9
google-auth==1.23.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
importlib-metadata==3.1.0
imutils==0.5.3
inflection==0.5.1
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
kiwisolver==1.3.1
kombu==5.0.2
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.4.0.42
opencv-python-headless==4.4.0.44
packaging==20.7
pandas==1.1.4
parso==0.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.8
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyDrive==1.3.1
Pygments==2.7.2
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzbar==0.1.8
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
traitlets==5.0.5
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
